css/css-images/parsing/object-position-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 3: parsing object-position with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-images-3/#propdef-object-position">
<meta name="assert" content="object-position supports the full '<position>' grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("object-position", "20% 30px");
test_valid_value("object-position", "10%", "10% center"); // "10% 50%" in Edge
test_valid_value("object-position", "right 30% top 60px"); // "calc(70%) 60px" in Edge.
test_valid_value("object-position", "left bottom"); // "0% 100%" in Edge. Similar for others below.
test_valid_value("object-position", "30px center");
test_valid_value("object-position", "40px top");
test_valid_value("object-position", "bottom 10% right 20%", "right 20% bottom 10%");
test_valid_value("object-position", "bottom right", "right bottom");
test_valid_value("object-position", "center 50px");
test_valid_value("object-position", "center bottom");
test_valid_value("object-position", "center left", "left center");
test_valid_value("object-position", "left", ["left center"]);
test_valid_value("object-position", "left center");
test_valid_value("object-position", "right 40%");
test_valid_value("object-position", "top", ["center top"]);
test_valid_value("object-position", "top center", "center top");

// ["center center"] in Blink and Firefox, "center" in WebKit. "50% 50%" in Edge.
test_valid_value("object-position", "center", ["center center", "center"]);
test_valid_value("object-position", "center center", ["center center", "center"]);
</script>
</body>
</html>
